requirements.txt
# Core ML packages
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
supervision>=0.1.0

# Geo-spatial packages
rasterio>=1.3.0
geopandas>=0.12.0
shapely>=2.0.0
pyproj>=3.5.0
folium>=0.14.0
rtree>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0
pyyaml>=6.0.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
jupyter>=1.0.0

# RF-DETR dependencies
git+https://github.com/IDEACVR/RF-Detr.git

# Optional: development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0